refactor(core): split legend builders by kind and shared helpers
Extract types, shared label/layout helpers, discrete, and continuous
(ramp+steps) builders from the monolithic legend-build module so each
kind can be edited without reviewing unrelated guide paths. Keep the
legend-build facade as the sole import surface for legend placement.
